Split up ContentBuilder settings component
