+++ front/app/components/admin/ContentBuilder/Settings/Settings.tsx
@@ -1,0 +1,83 @@
+import React from 'react';
+
+// styles
+import styled from 'styled-components';
+
+// components
+import {
+  Title,
+  Box,
+  stylingConsts,
+  colors,
+} from '@citizenlab/cl2-component-library';
+import Button from 'components/UI/Button';
+import CloseIconButton from 'components/UI/CloseIconButton';
+
+// intl
+import messages from '../messages';
+import { FormattedMessage } from 'utils/cl-intl';
+
+// typings
+import { Selected } from './typings';
+
+const StyledBox = styled(Box)`
+  box-shadow: -2px 0px 1px 0px rgba(0, 0, 0, 0.06);
+`;
+
+const StyledCloseIconButton = styled(CloseIconButton)`
+  position: absolute;
+  right: 8px;
+`;
+
+interface Props {
+  selected: Selected;
+  onClose: () => void;
+  onDelete: () => void;
+}
+
+const Settings = ({ selected, onClose, onDelete }: Props) => {
+  return (
+    <StyledBox
+      position="fixed"
+      right="0"
+      top={`${stylingConsts.menuHeight}px`}
+      zIndex="99999"
+      p="20px"
+      w="400px"
+      h={`calc(100vh - ${stylingConsts.menuHeight}px)`}
+      background="#ffffff"
+      overflowY="auto"
+    >
+      <StyledCloseIconButton
+        className="e2eBuilderSettingsClose"
+        a11y_buttonActionMessage={messages.a11y_closeSettingsPanel}
+        onClick={onClose}
+        iconColor={colors.textSecondary}
+        iconColorOnHover={'#000'}
+      />
+      {selected.title && (
+        <Title variant="h2">
+          <FormattedMessage {...selected.title} />
+        </Title>
+      )}
+      {selected.settings && React.createElement(selected.settings)}
+      {selected.isDeletable && !selected.custom?.noDelete ? (
+        <Box display="flex">
+          <Button
+            id="e2e-delete-button"
+            icon="delete"
+            buttonStyle="primary-outlined"
+            borderColor={colors.error}
+            textColor={colors.error}
+            iconColor={colors.error}
+            onClick={onDelete}
+          >
+            <FormattedMessage {...messages.delete} />
+          </Button>
+        </Box>
+      ) : null}
+    </StyledBox>
+  );
+};
+
+export default Settings;
