+++ front/app/components/admin/ContentBuilder/Settings/typings.ts
@@ -1,0 +1,12 @@
+import { MessageDescriptor } from 'utils/cl-intl';
+
+export interface Selected {
+  id: string;
+  name: string;
+  title: MessageDescriptor | undefined;
+  settings: React.ElementType<any> | undefined;
+  isDeletable: boolean;
+  custom?: {
+    noDelete?: boolean;
+  };
+}
